Add use of user and teamviewer install on Apps script
Adds lines for using the name of the user running the script to create
the downlod path for the packages used in the installation, instead of
using root to create that path, some computers did not resolve the user
name by themselves.
Adds lines for installation of teamviewer.
Changes the user for every package downloaded, changing from root to
user name.

	modified:   Fedora_26_27_Apps_Installation.sh