Improve the installation process
Adds a function to install every app and link sent from the script body,
it handles errors produced by the system if failing to install the
requested app.
Changes the link to the repo needed to install Fedy, since the old
version is not longer valid.
Also, it adds minor text changes to improve and make the script clearer
for reading.